requirements.txt
torch
torchvision
facenet-pytorch
onnxruntime
numpy
faiss-cpu
opencv-python
